Explore and Visualize Your Data Interactively

A 'shiny' gadget to create 'ggplot2' figures interactively with drag-and-drop to map your variables to different aesthetics. You can quickly visualize your data accordingly to their type, export in various formats, and retrieve the code to reproduce the plot.

See also

Author

Fanny Meyer & Victor Perrier (@dreamRs_fr)

Examples

if (FALSE) { # \dontrun{

esquisser()

# launch esquisse with specific data:
esquisser(mtcars)

} # }