An add-in to easily create plots with ggplot2

Select data to be used and map variables to aesthetics to produce a chart, customize common elements and get code to reproduce the chart.

Usage

esquisser(
  data = NULL,
  controls = c("options", "labs", "axes", "geoms", "theme", "filters", "code"),
  viewer = getOption(x = "esquisse.viewer", default = "dialog")
)

Arguments

data

a data.frame, you can pass a data.frame explicitly to the function, otherwise you'll have to choose one in global environment.

controls

Controls menu to be displayed. Use NULL to hide all menus.

viewer

Where to display the gadget: "dialog", "pane" or "browser" (see viewer).

Value

NULL. You can view code used to produce the chart, copy it or insert it in current script.

Examples

if (interactive()) {
# Launch with :
esquisser(iris)
# If in RStudio it will be launched by default in dialog window
# If not, it will be launched in browser

# Launch esquisse in browser :
esquisser(iris, viewer = "browser")

# You can set this option in .Rprofile :
options("esquisse.viewer" = "viewer")
# or
options("esquisse.viewer" = "browser")

# esquisse use shiny::runApp
# see ?shiny::runApp to see options
# available, example to use custom port:

options("shiny.port" = 8080)
esquisser(iris, viewer = "browser")

}