Drag And Drop Input Widget

Drag And Drop Input Widget

dragulaInput(inputId, sourceLabel, targetsLabels, targetsIds = NULL,
  choices = NULL, choiceNames = NULL, choiceValues = NULL,
  status = "primary", replace = FALSE, badge = TRUE, width = NULL,
  height = "200px")

Arguments

inputId
The input slot that will be used to access the value.

sourceLabel
Label display in the source box

targetsLabels
Labels for each target element.

targetsIds
Ids for retrieving values server-side, if NULL, the default, targetsLabels are used after removing all not-alphanumeric characters.

choices
List of values to select from (if elements of the list are named then that name rather than the value is displayed to the user). If this argument is provided, then choiceNames and choiceValues must not be provided, and vice-versa. The values should be strings; other types (such as logicals and numbers) will be coerced to strings.

choiceNames, choiceValues
List of names and values, respectively, that are displayed to the user in the app and correspond to the each choice (for this reason, choiceNames and choiceValues must have the same length). If either of these arguments is provided, then the other must be provided and choices must not be provided. The advantage of using both of these over a named list for choices is that choiceNames allows any type of UI object to be passed through (tag objects, icons, HTML code, ...), instead of just simple text.

status
If choices are displayed into a Bootstrap label, you can use Bootstrap status to color them, or NULL.

replace
When a choice is dragged in a target container already containing a choice, does the later be replaced by the new one ?

badge
Displays choices inside a Bootstrap badge. Use FALSE if you want to pass custom appearance with choiceNames.

width
Width of the input.

height
Height of each boxes, the total input height is this parameter X 2.

Value

a UI definition

Note

The output server-side is a list with two slots: source and targets.

See also

Examples

if (interactive()) {

library("shiny")
library("esquisse")

ui <- fluidPage(
  tags$h2("Demo dragulaInput"),
  tags$br(),
  dragulaInput(
    inputId = "dad",
    sourceLabel = "Source",
    targetsLabels = c("Target 1", "Target 2"),
    choices = names(iris),
    width = "400px"
  ),
  verbatimTextOutput(outputId = "result")
)

server <- function(input, output, session) {

  output$result <- renderPrint(str(input$dad))

}

shinyApp(ui = ui, server = server)

}