Match list of arguments to arguments of geometry

Match list of arguments to arguments of geometry

match_geom_args(
  geom,
  args,
  add_aes = TRUE,
  mapping = list(),
  envir = "ggplot2"
)

Arguments

geom
Character. name of the geometry.

args
Named list, parameters to be matched to the geometry arguments.

add_aes
Add aesthetics parameters (like size, fill, ...).

mapping
Mapping used in plot, to avoid setting fixed aesthetics parameters.

envir
Package environment to search in.

Value

a list

Examples

# List of parameters
params <- list(
  bins = 30,
  scale = "width",
  adjust = 2,
  position = "stack",
  size = 1.6,
  fill = "#112246"
)

# Search arguments according to geom
match_geom_args(geom = "histogram", args = params)

#> $bins
#> [1] 30
#> 
#> $fill
#> [1] "#112246"
#> 

match_geom_args(geom = "violin", args = params)

#> $scale
#> [1] "width"
#> 
#> $adjust
#> [1] 2
#> 
#> $fill
#> [1] "#112246"
#> 

match_geom_args(geom = "bar", args = params, add_aes = FALSE)

#> named list()

match_geom_args(geom = "point", args = params)

#> $size
#> [1] 1.6
#> 
#> $fill
#> [1] "#112246"
#> 

match_geom_args(geom = "point", args = params, add_aes = FALSE)

#> named list()