Module for choosing data.frame

DEPRECATED, please see package datamods for similar features.

Usage

chooseDataUI(id, label = "Data", icon = "database", width = "100%", ...)

chooseDataServer(
  input,
  output,
  session,
  dataModule = c("GlobalEnv", "ImportFile"),
  data = NULL,
  name = NULL,
  selectVars = TRUE,
  selectedTypes = c("continuous", "discrete", "time"),
  coerceVars = FALSE,
  launchOnStart = TRUE,
  size = "m"
)

Arguments

id

Module's id.

label

Label for button, passed to actionButton.

icon

Icon to appears on the button, passed to actionButton.

width

Width of button, passed to actionButton.

...

Other arguments passed to actionButton

input, output, session

standards shiny server arguments.

dataModule

Data module to use, choose between "GlobalEnv" (select ad data.frame from Global environment) or "ImportFile" (import an external file supported by import).

data

A data.frame to use by default.

name

Character, object's name to use for data.

selectVars

Display module to select variables, TRUE by default.

selectedTypes

Type of variables selected by default in select variables module. Possible types are "discrete", "time", "continuous" and "id", by default "id" is discarded.

coerceVars

Display module to coerce variables between different class, TRUE by default.

launchOnStart

Opens modal window when the application starts.

size

Size for the modal window.

Value

a reactiveValues containing the data selected under slot dataand the name of the selected data.frame under slot name.