Esquisse Shiny module

DEPRECATED, see esquisse-module.

esquisserServer(
  input,
  output,
  session,
  data = NULL,
  dataModule = c("GlobalEnv", "ImportFile"),
  sizeDataModule = "m"
)

esquisserUI(
  id,
  header = TRUE,
  container = esquisseContainer(),
  choose_data = TRUE,
  insert_code = FALSE,
  disable_filters = FALSE
)

Arguments

input, output, session
Standards shiny server arguments.

data
A reactiveValues with at least a slot data containing a data.frame to use in the module. And a slot name corresponding to the name of the data.frame.

dataModule
Data module to use, choose between "GlobalEnv" or "ImportFile".

sizeDataModule
Size for the modal window for selecting data.

id
Module's id.

header
Logical. Display or not esquisse header.

container
Container in which display the addin, default is to use esquisseContainer, see examples. Use NULL for no container (behavior in versions <= 0.2.1). Must be a function.

choose_data
Logical. Display or not the button to choose data.

insert_code
Logical, Display or not a button to insert the ggplot code in the current user script (work only in RStudio).

disable_filters
Logical. Disable the menu allowing to filter data used.

Value

A reactiveValues with 3 slots :

code_plot : code to generate plot.

code_filters : a list of length two with code to reproduce filters.

data : data.frame used in plot (with filters applied).

Note

For the module to display correctly, it is necessary to place it in a container with a fixed height. Since version >= 0.2.2, the container is added by default.