Shiny module to interactively filter a data.frame

DEPRECATED, please see package datamods for similar features.

filterDF_UI(id, show_nrow = TRUE)

filterDF(
  input,
  output,
  session,
  data_table = reactive(),
  data_vars = shiny::reactive(NULL),
  data_name = reactive("data"),
  label_nrow = "Number of rows:",
  drop_ids = TRUE,
  picker = FALSE
)

Arguments

id
Module id. See callModule.

show_nrow
Show number of filtered rows and total.

input, output, session
standards shiny server arguments.

data_table
reactive function returning a data.frame to filter.

data_vars
reactive function returning a character vector of variable to use for filters.

data_name
reactive function returning a character string representing data_table name.

label_nrow
Text to display before the number of rows of filtered data / source data.

drop_ids
Drop columns containing more than 90% of unique values, or than 50 distinct values.

picker
Use shinyWidgets::pickerInput instead of shiny::selectizeInput (default).

Value

A list with 2 elements :

data_filtered : reactive function returning data filtered.

code : reactiveValues with 2 slots : expr (raw expression to filter data) and dplyr (code with dplyr pipeline).